requirements.txt
asgiref==3.6.0
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.0.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.2.0
Django==4.1.5
django-import-export==3.0.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.4.0
psycopg2-binary==2.9.5
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
sqlparse==0.4.3
tablib==3.3.0
tzdata==2022.7
urllib3==1.26.14
whitenoise==6.3.0
xlrd==2.0.1
xlwt==1.3.0
